+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Version listings include the following sections, if applicable:
 
 See also <https://github.com/xgi/castero/releases>.
 
-## Unreleased
+## 0.2.6 - 2018-08-15
 **Added**
 * Added support for custom keybindings in the config.